feat(poll): re-add import aliases that were removed by IDE refactoring
