+++ tests/integration/server/finance/account.repository/findByUserId.test.ts
@@ -1,6 +1,6 @@
-import AccountRepository from "server/finance/repositories/account.repository";
-import { expect, test } from "test-utils/vitest";
-import { anyId } from "test-utils/vitest/patterns";
+import AccountRepository from "~~/server/finance/repositories/account.repository";
+import { expect, test } from "~~/test-utils/vitest";
+import { anyId } from "~~/test-utils/vitest/patterns";
 
 test("returns only the accounts belonging to the requested user", async ({ db, user }) => {
     const [otherUser] = await db.user.insert(1, { name: "Smithers", email: "smithers@burns.com" });
