+++ tests/integration/server/finance/account.repository/update.test.ts
@@ -1,6 +1,6 @@
-import AccountRepository from "server/finance/repositories/account.repository";
-import { BASE_ACCOUNT } from "test-utils/constants/finance";
-import { expect, test } from "test-utils/vitest";
+import AccountRepository from "~~/server/finance/repositories/account.repository";
+import { BASE_ACCOUNT } from "~~/test-utils/constants/finance";
+import { expect, test } from "~~/test-utils/vitest";
 
 let id = "";
 test.beforeEach(async ({ db, user }) => {
