+++ tests/integration/server/finance/auto-transaction.repository/findByUserAndAccountId.test.ts
@@ -1,6 +1,6 @@
-import AutoTransactionRepository from "server/finance/repositories/auto-transaction.repository";
-import { expect, test } from "test-utils/vitest";
-import { anyId } from "test-utils/vitest/patterns";
+import AutoTransactionRepository from "~~/server/finance/repositories/auto-transaction.repository";
+import { expect, test } from "~~/test-utils/vitest";
+import { anyId } from "~~/test-utils/vitest/patterns";
 
 test("returns only the accounts belonging to the requested user", async ({ db, user }) => {
     const [otherUser] = await db.user.insert(1, { name: "Smithers", email: "smithers@burns.com" });
