+++ tests/integration/server/finance/auto-transaction.repository/update.test.ts
@@ -1,6 +1,6 @@
-import AutoTransactionRepository from "server/finance/repositories/auto-transaction.repository";
-import { BASE_AUTO_TRANSACTION } from "test-utils/constants/finance";
-import { expect, test } from "test-utils/vitest";
+import AutoTransactionRepository from "~~/server/finance/repositories/auto-transaction.repository";
+import { BASE_AUTO_TRANSACTION } from "~~/test-utils/constants/finance";
+import { expect, test } from "~~/test-utils/vitest";
 
 let id = "";
 test.beforeEach(async ({ db, user }) => {
