+++ tests/integration/server/poll/poll.repository/findByUserId.test.ts
@@ -1,6 +1,6 @@
-import PollRepository from "server/poll/repositories/poll.repository";
-import { expect, test } from "test-utils/vitest";
-import { anyId } from "test-utils/vitest/patterns";
+import PollRepository from "~~/server/poll/repositories/poll.repository";
+import { expect, test } from "~~/test-utils/vitest";
+import { anyId } from "~~/test-utils/vitest/patterns";
 
 test("returns only the polls belonging to the requested user", async ({ db, user }) => {
     const [otherUser] = await db.user.insert(1, { name: "Smithers", email: "smithers@burns.com" });
