+++ tests/integration/server/timer/timer.repository/create.test.ts
@@ -1,6 +1,6 @@
-import TimerRepository from "server/timer/repositories/timer.repository";
-import { BASE_TIMER } from "test-utils/constants/timer";
-import { expect, test } from "test-utils/vitest";
+import TimerRepository from "~~/server/timer/repositories/timer.repository";
+import { BASE_TIMER } from "~~/test-utils/constants/timer";
+import { expect, test } from "~~/test-utils/vitest";
 
 test("creates a timer with intervals and returns the id of the created timer", async ({ db, user }) => {
     const timerRepository = new TimerRepository(db.client);
