+++ tests/integration/server/timer/timer.repository/update.test.ts
@@ -1,7 +1,7 @@
-import TimerRepository from "server/timer/repositories/timer.repository";
-import { BASE_TIMER } from "test-utils/constants/timer";
-import { expect, test } from "test-utils/vitest";
-import { anyId } from "test-utils/vitest/patterns";
+import TimerRepository from "~~/server/timer/repositories/timer.repository";
+import { BASE_TIMER } from "~~/test-utils/constants/timer";
+import { expect, test } from "~~/test-utils/vitest";
+import { anyId } from "~~/test-utils/vitest/patterns";
 
 let id = "";
 test.beforeEach(async ({ db, user }) => {
